Update dev dependencies
codecov

add @vitest/coverage-v8

format

vite.config.mjs

fix tests

build
