+++ .github/workflows/ci.yml
@@ -5,21 +5,32 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        node-version: [14.x, 16.x]
+        node-version: [18.x, 20.x, 22.x]
     steps:
-    - name: Checkout
-      uses: actions/checkout@v2
-    - name: Use Node.js ${{ matrix.node-version }}
-      uses: actions/setup-node@v2
-      with:
-        node-version: ${{ matrix.node-version }}
-    - name: Install dependencies
-      run: npm ci --force
-    - name: Lint
-      run: npm run lint
-    - name: Test
-      run: npm run test -- --coverage --verbose --ci
-    - name: Build
-      run: npm run build
-    - name: Upload coverage to Codecov
-      uses: codecov/codecov-action@v1
+      - name: Checkout
+        uses: actions/checkout@v4
+      - name: Use Node.js ${{ matrix.node-version }}
+        uses: actions/setup-node@v4
+        with:
+          node-version: ${{ matrix.node-version }}
+      - name: Install dependencies
+        run: npm ci
+      - name: Run Tests
+        run: npm run coverage -- --reporter=junit --outputFile=test-report.junit.xml
+      - name: Lint code
+        run: npm run lint
+      - name: Check types
+        run: npm run typecheck
+      - name: Check code formatting
+        run: npm run formatcheck
+      - name: Build
+        run: npm run build
+      - name: Upload coverage to Codecov
+        uses: codecov/codecov-action@v4
+        with:
+          token: ${{ secrets.CODECOV_TOKEN }}
+      - name: Upload test results to Codecov
+        if: ${{ !cancelled() }}
+        uses: codecov/test-results-action@v1
+        with:
+          token: ${{ secrets.CODECOV_TOKEN }}
