+++ .gitignore
@@ -2,19 +2,16 @@
 
 # Dependencies
 node_modules/
-yarn.lock
+src/path-to-regexp.ts
 
 # Compiled output
-/dist/*
-!/dist/universal-router*
+/dist/
 
 # Test coverage
 /coverage/
 
 # Logs
 npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
 
 # Editors and IDEs
 .idea/
