--- .prettierrc.js
@@ -1,7 +1,0 @@
-module.exports = {
-  printWidth: 100,
-  semi: false,
-  singleQuote: true,
-  trailingComma: 'all',
-  arrowParens: 'always',
-}
