+++ docs/getting-started.md
@@ -22,12 +22,12 @@ import UniversalRouter from 'universal-router'
 const routes = [
   { path: '/one', action: () => '<h1>Page One</h1>' },
   { path: '/two', action: () => '<h1>Page Two</h1>' },
-  { path: '/*all', action: () => '<h1>Not Found</h1>' }
+  { path: '/*all', action: () => '<h1>Not Found</h1>' },
 ]
 
 const router = new UniversalRouter(routes)
 
-router.resolve({ pathname: '/one' }).then(result => {
+router.resolve({ pathname: '/one' }).then((result) => {
   document.body.innerHTML = result
   // renders: <h1>Page One</h1>
 })
@@ -61,12 +61,12 @@ import UniversalRouter from 'universal-router'
 const routes = [
   { path: '/one', action: () => <h1>Page One</h1> },
   { path: '/two', action: () => <h1>Page Two</h1> },
-  { path: '/*all', action: () => <h1>Not Found</h1> }
+  { path: '/*all', action: () => <h1>Not Found</h1> },
 ]
 
 const router = new UniversalRouter(routes)
 
-router.resolve({ pathname: '/one' }).then(component => {
+router.resolve({ pathname: '/one' }).then((component) => {
   ReactDOM.render(component, document.body)
   // renders: <h1>Page One</h1>
 })
