+++ eslint.config.mjs
@@ -1,0 +1,66 @@
+import js from '@eslint/js'
+import importPlugin from 'eslint-plugin-import'
+import simpleImportSort from 'eslint-plugin-simple-import-sort'
+import { config, configs } from 'typescript-eslint'
+import globals from 'globals'
+
+export default config(
+  {
+    ignores: [
+      '**/node_modules/',
+      '**/dist/',
+      '**/build/',
+      '**/coverage/',
+      'src/path-to-regexp.ts',
+    ],
+  },
+  js.configs.recommended,
+  ...configs.recommendedTypeChecked,
+  ...configs.stylisticTypeChecked,
+  // eslint-disable-next-line @typescript-eslint/no-unsafe-argument, @typescript-eslint/no-unsafe-member-access
+  importPlugin.flatConfigs.recommended,
+  // eslint-disable-next-line @typescript-eslint/no-unsafe-argument, @typescript-eslint/no-unsafe-member-access
+  importPlugin.flatConfigs.typescript,
+  {
+    languageOptions: {
+      parserOptions: {
+        projectService: {
+          allowDefaultProject: ['*.ts', '*.js', '*.mjs', 'tools/*.js'],
+        },
+        tsconfigRootDir: import.meta.name,
+      },
+    },
+    settings: {
+      'import/resolver': {
+        node: true,
+        typescript: { alwaysTryTypes: true, project: ['./tsconfig.json'] },
+      },
+    },
+    plugins: { 'simple-import-sort': simpleImportSort },
+    rules: {
+      '@typescript-eslint/consistent-indexed-object-style': 'off',
+      '@typescript-eslint/consistent-type-definitions': 'off',
+      '@typescript-eslint/no-explicit-any': 'off',
+      '@typescript-eslint/no-unsafe-assignment': 'off',
+      '@typescript-eslint/prefer-for-of': 'off',
+      '@typescript-eslint/prefer-nullish-coalescing': 'off',
+      '@typescript-eslint/prefer-string-starts-ends-with': 'off',
+    },
+  },
+  {
+    files: ['src/**/*.test.ts'],
+    rules: {
+      '@typescript-eslint/no-unsafe-member-access': 'off',
+      '@typescript-eslint/no-unsafe-return': 'off',
+    },
+  },
+  {
+    files: ['tools/**/*.js'],
+    languageOptions: { globals: { ...globals.node } },
+    rules: {
+      '@typescript-eslint/no-require-imports': 'off',
+      '@typescript-eslint/no-unsafe-argument': 'off',
+      '@typescript-eslint/no-unused-vars': 'off',
+    },
+  },
+)
