--- jest.config.js
@@ -1,5 +1,0 @@
-// https://jestjs.io/docs/en/configuration
-module.exports = {
-  preset: 'ts-jest',
-  roots: ['<rootDir>/src/'],
-}
