+++ prettier.config.mjs
@@ -1,0 +1,10 @@
+/**
+ * @see https://prettier.io/docs/en/configuration.html
+ * @type {import("prettier").Config}
+ */
+const config = {
+  semi: false,
+  singleQuote: true,
+}
+
+export default config
