+++ src/universal-router-sync.ts
@@ -7,16 +7,16 @@
  * LICENSE.txt file in the root directory of this source tree.
  */
 
-import {
-  match,
+import { match } from './path-to-regexp.js'
+import type {
   Path,
   Match,
   MatchFunction,
   ParseOptions,
   MatchOptions,
   PathToRegexpOptions,
-  CompileOptions
-} from 'path-to-regexp'
+  CompileOptions,
+} from './path-to-regexp.js'
 
 /**
  * In addition to a URL path string, any arbitrary data can be passed to
@@ -97,7 +97,10 @@ export interface Route<R = any, C extends RouterContext = RouterContext> {
    * Action method should return anything except `null` or `undefined` to be resolved by router
    * otherwise router will throw `Page not found` error if all matched routes returned nothing.
    */
-  action?: (context: RouteContext<R, C>, params: RouteParams) => RouteResultSync<R>
+  action?: (
+    context: RouteContext<R, C>,
+    params: RouteParams,
+  ) => RouteResultSync<R>
   /**
    * The route path match function. Used for internal caching.
    */
@@ -109,7 +112,10 @@ export interface Route<R = any, C extends RouterContext = RouterContext> {
  * @template C User context that is made union with RouterContext.
  * @template R Result that every action function resolves to.
  */
-export type Routes<R = any, C extends RouterContext = RouterContext> = Array<Route<R, C>>
+export type Routes<R = any, C extends RouterContext = RouterContext> = Route<
+  R,
+  C
+>[]
 
 export type ResolveRoute<R = any, C extends RouterContext = RouterContext> = (
   context: RouteContext<R, C>,
@@ -144,7 +150,7 @@ export interface RouteMatch<R = any, C extends RouterContext = RouterContext> {
 function decode(val: string): string {
   try {
     return decodeURIComponent(val)
-  } catch (err) {
+  } catch {
     return val
   }
 }
@@ -157,11 +163,17 @@ function matchRoute<R, C extends RouterContext>(
   parentParams?: RouteParams,
 ): Iterator<RouteMatch<R, C>, false, Route<R, C> | false> {
   let matchResult: Match<RouteParams>
-  let childMatches: Iterator<RouteMatch<R, C>, false, Route<R, C> | false> | null
+  let childMatches: Iterator<
+    RouteMatch<R, C>,
+    false,
+    Route<R, C> | false
+  > | null
   let childIndex = 0
 
   return {
-    next(routeToSkip: Route<R, C> | false): IteratorResult<RouteMatch<R, C>, false> {
+    next(
+      routeToSkip: Route<R, C> | false,
+    ): IteratorResult<RouteMatch<R, C>, false> {
       if (route === routeToSkip) {
         return { done: true, value: false }
       }
@@ -176,7 +188,8 @@ function matchRoute<R, C extends RouterContext>(
 
         if (matchResult) {
           const { path } = matchResult
-          matchResult.path = !end && path.charAt(path.length - 1) === '/' ? path.substr(1) : path
+          matchResult.path =
+            !end && path.charAt(path.length - 1) === '/' ? path.substr(1) : path
           matchResult.params = { ...parentParams, ...matchResult.params }
           return {
             done: false,
@@ -207,10 +220,7 @@ function matchRoute<R, C extends RouterContext>(
 
           const childMatch = childMatches.next(routeToSkip)
           if (!childMatch.done) {
-            return {
-              done: false,
-              value: childMatch.value,
-            }
+            return { done: false, value: childMatch.value }
           }
 
           childMatches = null
@@ -254,14 +264,19 @@ class UniversalRouterSync<R = any, C extends RouterContext = RouterContext> {
 
   options: RouterOptions<R, C>
 
-  constructor(routes: Routes<R, C> | Route<R, C>, options?: RouterOptions<R, C>) {
+  constructor(
+    routes: Routes<R, C> | Route<R, C>,
+    options?: RouterOptions<R, C>,
+  ) {
     if (!routes || typeof routes !== 'object') {
       throw new TypeError('Invalid routes')
     }
 
     this.options = { decode, ...options }
     this.baseUrl = this.options.baseUrl || ''
-    this.root = Array.isArray(routes) ? { path: '', children: routes, parent: null } : routes
+    this.root = Array.isArray(routes)
+      ? { path: '', children: routes, parent: null }
+      : routes
     this.root.parent = null
   }
 
@@ -294,7 +309,8 @@ class UniversalRouterSync<R = any, C extends RouterContext = RouterContext> {
       parent: Route<R, C> | false = !matches.done && matches.value.route,
       prevResult?: RouteResultSync<R>,
     ): RouteResultSync<R> {
-      const routeToSkip = prevResult === null && !matches.done && matches.value.route
+      const routeToSkip =
+        prevResult === null && !matches.done && matches.value.route
       matches = nextMatches || matchResult.next(routeToSkip)
       nextMatches = null
 
@@ -313,7 +329,10 @@ class UniversalRouterSync<R = any, C extends RouterContext = RouterContext> {
 
       currentContext = { ...context, ...matches.value }
 
-      const result = resolve(currentContext as RouteContext<R, C>, matches.value.params)
+      const result = resolve(
+        currentContext as RouteContext<R, C>,
+        matches.value.params,
+      )
       if (result !== null && result !== undefined) {
         return result
       }
