+++ src/universal-router.ts
@@ -7,16 +7,16 @@
  * LICENSE.txt file in the root directory of this source tree.
  */
 
-import {
-  match,
+import { match } from './path-to-regexp.js'
+import type {
   Path,
   Match,
   MatchFunction,
   ParseOptions,
   MatchOptions,
   PathToRegexpOptions,
-  CompileOptions
-} from 'path-to-regexp'
+  CompileOptions,
+} from './path-to-regexp.js'
 
 /**
  * In addition to a URL path string, any arbitrary data can be passed to
@@ -40,7 +40,11 @@ export interface RouteParams {
   [paramName: string]: string | string[]
 }
 
-export type RouteResult<T> = T | null | undefined | Promise<T | null | undefined>
+export type RouteResult<T> =
+  | T
+  | null
+  | undefined
+  | Promise<T | null | undefined>
 
 export interface RouteContext<R = any, C extends RouterContext = RouterContext>
   extends ResolveContext {
@@ -111,7 +115,10 @@ export interface Route<R = any, C extends RouterContext = RouterContext> {
  * @template R Result that every action function resolves to.
  * If the action returns a Promise, R can be the type the Promise resolves to.
  */
-export type Routes<R = any, C extends RouterContext = RouterContext> = Array<Route<R, C>>
+export type Routes<R = any, C extends RouterContext = RouterContext> = Route<
+  R,
+  C
+>[]
 
 export type ResolveRoute<R = any, C extends RouterContext = RouterContext> = (
   context: RouteContext<R, C>,
@@ -120,7 +127,10 @@ export type ResolveRoute<R = any, C extends RouterContext = RouterContext> = (
 
 export type RouteError = Error & { status?: number }
 
-export type ErrorHandler<R = any> = (error: RouteError, context: ResolveContext) => RouteResult<R>
+export type ErrorHandler<R = any> = (
+  error: RouteError,
+  context: ResolveContext,
+) => RouteResult<R>
 
 export interface RouterOptions<R = any, C extends RouterContext = RouterContext>
   extends ParseOptions,
@@ -143,7 +153,7 @@ export interface RouteMatch<R = any, C extends RouterContext = RouterContext> {
 function decode(val: string): string {
   try {
     return decodeURIComponent(val)
-  } catch (err) {
+  } catch {
     return val
   }
 }
@@ -156,11 +166,17 @@ function matchRoute<R, C extends RouterContext>(
   parentParams?: RouteParams,
 ): Iterator<RouteMatch<R, C>, false, Route<R, C> | false> {
   let matchResult: Match<RouteParams>
-  let childMatches: Iterator<RouteMatch<R, C>, false, Route<R, C> | false> | null
+  let childMatches: Iterator<
+    RouteMatch<R, C>,
+    false,
+    Route<R, C> | false
+  > | null
   let childIndex = 0
 
   return {
-    next(routeToSkip: Route<R, C> | false): IteratorResult<RouteMatch<R, C>, false> {
+    next(
+      routeToSkip: Route<R, C> | false,
+    ): IteratorResult<RouteMatch<R, C>, false> {
       if (route === routeToSkip) {
         return { done: true, value: false }
       }
@@ -175,7 +191,8 @@ function matchRoute<R, C extends RouterContext>(
 
         if (matchResult) {
           const { path } = matchResult
-          matchResult.path = !end && path.charAt(path.length - 1) === '/' ? path.substr(1) : path
+          matchResult.path =
+            !end && path.charAt(path.length - 1) === '/' ? path.substr(1) : path
           matchResult.params = { ...parentParams, ...matchResult.params }
           return {
             done: false,
@@ -206,10 +223,7 @@ function matchRoute<R, C extends RouterContext>(
 
           const childMatch = childMatches.next(routeToSkip)
           if (!childMatch.done) {
-            return {
-              done: false,
-              value: childMatch.value,
-            }
+            return { done: false, value: childMatch.value }
           }
 
           childMatches = null
@@ -253,14 +267,19 @@ class UniversalRouter<R = any, C extends RouterContext = RouterContext> {
 
   options: RouterOptions<R, C>
 
-  constructor(routes: Routes<R, C> | Route<R, C>, options?: RouterOptions<R, C>) {
+  constructor(
+    routes: Routes<R, C> | Route<R, C>,
+    options?: RouterOptions<R, C>,
+  ) {
     if (!routes || typeof routes !== 'object') {
       throw new TypeError('Invalid routes')
     }
 
     this.options = { decode, ...options }
     this.baseUrl = this.options.baseUrl || ''
-    this.root = Array.isArray(routes) ? { path: '', children: routes, parent: null } : routes
+    this.root = Array.isArray(routes)
+      ? { path: '', children: routes, parent: null }
+      : routes
     this.root.parent = null
   }
 
@@ -293,7 +312,8 @@ class UniversalRouter<R = any, C extends RouterContext = RouterContext> {
       parent: Route<R, C> | false = !matches.done && matches.value.route,
       prevResult?: RouteResult<R>,
     ): Promise<RouteResult<R>> {
-      const routeToSkip = prevResult === null && !matches.done && matches.value.route
+      const routeToSkip =
+        prevResult === null && !matches.done && matches.value.route
       matches = nextMatches || matchResult.next(routeToSkip)
       nextMatches = null
 
@@ -326,7 +346,7 @@ class UniversalRouter<R = any, C extends RouterContext = RouterContext> {
 
     return Promise.resolve()
       .then(() => next(true, this.root))
-      .catch((error) => {
+      .catch((error: RouteError) => {
         if (this.options.errorHandler) {
           return this.options.errorHandler(error, currentContext)
         }
