--- tools/.eslintrc.js
@@ -1,12 +1,0 @@
-module.exports = {
-  rules: {
-    'import/no-extraneous-dependencies': [
-      'error',
-      {
-        optionalDependencies: false,
-        peerDependencies: false,
-      },
-    ],
-    'no-console': 'off',
-  },
-}
