--- tools/build.tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "module": "es6",
-    "rootDir": "../dist/src"
-  },
-  "include": ["../dist/src"]
-}
