+++ tools/tsconfig.esm.json
@@ -1,0 +1,4 @@
+{
+  "extends": "../tsconfig.json",
+  "exclude": ["../src/**/*.test.ts"]
+}
