+++ vite.config.mjs
@@ -1,0 +1,11 @@
+/// <reference types="vitest" />
+import { defineConfig } from 'vite'
+
+export default defineConfig({
+  test: {
+    coverage: {
+      include: ['src'],
+      exclude: ['src/path-to-regexp.ts'],
+    },
+  },
+})
